Meta files added
Unity requires an option to be turned on in order to keep important
settings about mappings and stuff. This was turned off obviously and I
had to map everything again. Also when we turn that setting on we need
to explicitly allow checking in of ".meta" files which I added now.